Fix TypeError in PubSubClass.unsubscribe
Modifying an array while iterating over it is tricky.  If we delete an
element in the middle and continue to iterate using the old length,
we'll perform an access beyond the end of the array, get back an
undefined, and then get a TypeError when we try to [0] it.

I don't know if this method was meant to remove just the first matching
callback, or all of them.  My modification assumes the first; if you
need the second, instead of break do `i--; l--`.